fix: make interceptor shutdown retriable
