+++ docs/README/ja.md
@@ -110,7 +110,7 @@ interceptor.start();
 | メソッド | 説明 |
 | --- | --- |
 | `start()` | `fetch` と `XMLHttpRequest` を上書きして傍受を開始します。導入に失敗した場合は完了済みの変更を戻し、停止状態を維持します。 |
-| `stop()` | 傍受を停止します。一部の復元に失敗しても、必要な復元をすべて試みます。 |
+| `stop()` | 元のブラウザ API をすべて復元します。一部の復元に失敗した場合は例外を投げ、アダプターを重ねずに再度 `stop()` で失敗分を再試行できる状態を維持します。 |
 
 ## ユースケース
