+++ docs/README/ko.md
@@ -110,7 +110,7 @@ interceptor.start();
 | 메서드 | 설명 |
 | --- | --- |
 | `start()` | `fetch` 와 `XMLHttpRequest` 를 override 하여 인터셉션을 시작합니다. 설치가 실패하면 완료된 변경을 되돌리고 중지 상태를 유지합니다. |
-| `stop()` | 인터셉션을 중지합니다. 일부 복원이 실패해도 필요한 모든 복원을 시도합니다. |
+| `stop()` | 원래 브라우저 API 를 모두 복원합니다. 일부 복원이 실패하면 예외를 던지고, 어댑터를 중첩하지 않은 채 다시 `stop()` 을 호출해 실패한 복원을 재시도할 수 있는 상태를 유지합니다. |
 
 ## 사용 사례
