+++ docs/README/zh-CN.md
@@ -110,7 +110,7 @@ interceptor.start();
 | 方法 | 说明 |
 | --- | --- |
 | `start()` | 覆盖 `fetch` 和 `XMLHttpRequest` 以开始拦截。安装失败时会回滚已完成的修改并保持停止状态。 |
-| `stop()` | 停止拦截。即使某一步恢复失败，也会尝试所有必要的恢复操作。 |
+| `stop()` | 尝试恢复所有原始浏览器 API。若部分恢复失败，它会抛出异常并保留可重试状态，可再次调用 `stop()`，且不会叠加新的适配器。 |
 
 ## 使用场景
